Don't refer app, require state ns :as state
:refering app into the namespace means that the var
will always be resolved. I ran into an error where I wrote
a function that was supposed to get `app` as a parameter, but it didn't,
and I didn't notice because the missing argument wasn't highlighted,
since my editor thought I meant the app that was required on the top
level of the namespace.

Let's remove this beginner trap :)